refactor(Services): ♻️ Decouple `scanner` from `api_config`.
Seperate the conserns for `scanner_api` & `api_config`.
